update types, pass activations, split history/activity into separate files
