+++ static/app/views/alerts/rules/metric/types.tsx
@@ -3,7 +3,7 @@ import type {ActivationConditionType, MonitorType} from 'sentry/types/alerts';
 import type {MEPAlertsQueryType} from 'sentry/views/alerts/wizard/options';
 import type {SchemaFormConfig} from 'sentry/views/settings/organizationIntegrations/sentryAppExternalForm';
 
-import type {Incident} from '../../types';
+import type {AlertRuleActivation, Incident} from '../../types';
 
 export enum AlertRuleThresholdType {
   ABOVE = 0,
@@ -84,16 +84,6 @@ export type SavedTrigger = Omit<UnsavedTrigger, 'actions'> & {
 };
 
 export type Trigger = Partial<SavedTrigger> & UnsavedTrigger;
-
-export type AlertRuleActivation = {
-  alertRuleId: string;
-  dateCreated: string;
-  finishedAt: string;
-  id: string;
-  isComplete: boolean;
-  querySubscriptionId: string;
-  metricValue?: number;
-};
 
 export enum ActivationCondition {
   RELEASE_CONDITION = 0,
